Add NestedCollectionMixin and nested activity collections for habits
Querying habit/<id>/activity/ should return activities belonging to that
particular Habit. All of the usual list/crud ops should be possible. To achieve
this, a mixin hooks the get_queryset method used by DRF before performing the
list/crud ops. This filters the queryset and allows for 'pk' or 'pk1' when
there are multiple URL args for pk.

The view has to supply this mixin with info about the parent collection field
name (e.g. 'habit' in this case) and if 'pk' is no longer the name of the URL
arg to lookup the child resource, 'lookup_url_kwarg' should be defined too.